data-fetcher/requirements.txt
ccxt
requests
pandas
numpy
scikit-learn
tensorflow
kafka-python
